Add files via upload
Simple laundry application developed using Netbeans (IDE). The language used to write this application is Java with the use of JForms. The application is developed with the intention to simplify mechanical operations at a local laundromat that used charts and scales with receipts written out on paper. The full development of this application would include a database to store information about resources at the user's disposable and track usage. This would simplify the process and optimise the business process.